Move keyboard controls to video controls
